Initialize gradle project structure
