+++ .gitignore
@@ -1,0 +1,20 @@
+.DS_STORE
+.gradle/
+build/
+*.iml
+*.idea
+*/.classpath
+*/.project
+*/.settings
+*/bin
+*/*.jar
+*/lib/spigot*.jar
+*/libs/spigot*.jar
+*/dist/lib/spigot*.jar
+*/build
+*/nbproject
+*.prefs
+*.class
+.project
+/bin
+idea/*
