+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Jan 05 14:25:12 CET 2020
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.0.1-all.zip
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
