+++ src/main/java/com/thizthizzydizzy/treefeller/GriefPreventionCompat.java
@@ -1,0 +1,13 @@
+package com.thizthizzydizzy.treefeller;
+import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
+public class GriefPreventionCompat extends InternalCompatibility{
+    @Override
+    public String getPluginName(){
+        return "GriefPrevention";
+    }
+    @Override
+    public boolean test(Player player, Block block) {
+        return me.ryanhamshire.GriefPrevention.GriefPrevention.instance.allowBreak(player, block, block.getLocation())!=null;
+    }
+}
