+++ src/main/java/com/thizthizzydizzy/treefeller/InternalCompatibility.java
